+++ sdk/education/arm-education/CHANGELOG.md
@@ -1,15 +1,41 @@
 # Release History
+    
+## 1.0.0-beta.3 (2022-10-11)
+    
+**Features**
 
-## 1.0.0-beta.3 (Unreleased)
+  - Added operation group Grant
+  - Added operation group OperationOperations
+  - Added operation Grants.getV2
+  - Added operation Grants.listAllV2
+  - Added operation Grants.listV2
+  - Added Interface GrantDetailsV2
+  - Added Interface GrantListResponseV2
+  - Added Interface GrantRenewalHeaders
+  - Added Interface GrantRenewalOptionalParams
+  - Added Interface GrantsGetV2OptionalParams
+  - Added Interface GrantsListAllV2NextOptionalParams
+  - Added Interface GrantsListAllV2OptionalParams
+  - Added Interface GrantsListV2NextOptionalParams
+  - Added Interface GrantsListV2OptionalParams
+  - Added Interface OperationStatusHeaders
+  - Added Interface OperationStatusOptionalParams
+  - Added Type Alias GrantRenewalResponse
+  - Added Type Alias GrantsGetV2Response
+  - Added Type Alias GrantsListAllV2NextResponse
+  - Added Type Alias GrantsListAllV2Response
+  - Added Type Alias GrantsListV2NextResponse
+  - Added Type Alias GrantsListV2Response
+  - Added Type Alias OperationStatusResponse
+  - Class EducationManagementClient has a new parameter grant
+  - Class EducationManagementClient has a new parameter operationOperations
 
-### Features Added
+**Breaking Changes**
 
-### Breaking Changes
-
-### Bugs Fixed
-
-### Other Changes
-
+  - Interface EducationManagementClientOptionalParams no longer has parameter apiVersion
+  - Class EducationManagementClient no longer has parameter apiVersion
+    
+    
 ## 1.0.0-beta.2 (2022-07-04)
     
 **Features**